refactor: extract pending resolution helpers and single-flight reconnect
